requirements[1].txt
aiogram==3.3.0
gspread==6.0.0
google-auth==2.29.0
python-dotenv==1.0.1
